students/k3341/Skoblilova_Viktoria/Lr1/gateway/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "alembic>=1.16.1",
    "apscheduler>=3.10.4", 
    "asyncpg>=0.30.0",
    "audiofile>=1.5.1",
    "bcrypt>=4.1.2",
    "fastapi>=0.45.0",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "passlib[argon2]>=1.7.4",
    "polars>=1.31.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "scalar-fastapi>=1.0.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "xlsxwriter>=3.2.5",
    "aiosmtplib>=3.0.0", 
    "email-validator>=2.0.0",
]

[dependency-groups]
dev = [
    "jedi>=0.19.2",
    "jedi-language-server>=0.45.1",
    "pyrefly>=0.17.0",
    "ruff>=0.11.11",
]

[tool.pyrefly]
ignore-errors-in-generated-code = true
use-untyped-imports = true
disableLanguageServices=false 
disableTypeErrors=true

[tool.pyrefly.errors]
bad-function-definition = false

[tool.ruff.lint]
select = ['E', 'F', 'UP', 'B', 'SIM', 'I']
ignore = ['B008']

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
